move the data subdirectories to the top level of the rank app
