--- rank/data/README.md
@@ -1,13 +1,0 @@
-# Data
-
-## indices
-
-where mappings are stored
-
-## terms
-
-real lists of search terms collected from the reporting cluster
-
-## tests
-
-the rank_eval tests we run to make sure that queries perform as expected
