--- rank/data/languages.ts
@@ -1,1 +1,0 @@
-export default ['french', 'german', 'italian', 'hindi', 'bengali', 'arabic']
