--- rank/data/tests/pass.ts
@@ -1,26 +1,0 @@
-import { estypes } from '@elastic/elasticsearch'
-
-type RankDetail = estypes.RankEvalRankEvalMetricDetail
-
-type Pass = {
-  score: number // should be a range between 0 => 1
-  pass: boolean
-}
-type Eval = (score: RankDetail) => Pass
-
-const equalTo1: Eval = (rankDetail: RankDetail) => {
-  return {
-    score: rankDetail.metric_score,
-    pass: rankDetail.metric_score === 1,
-  }
-}
-
-const equalTo0: Eval = (rankDetail: RankDetail) => {
-  return {
-    score: rankDetail.metric_score,
-    pass: rankDetail.metric_score === 0,
-  }
-}
-
-export { equalTo0, equalTo1 }
-export type { Pass, Eval }
